[FEAT] Mission2 관련 D2-D5 관련 추가
1. Display 관련 Factoring Pattern 적용.
2. TDD 관련 Unit Test 기본 추가
