+++ mission2/mission2/assemble.cpp
@@ -1,5 +1,5 @@
-﻿#ifdef _DEBUG
-
+﻿#ifdef NDEBUG
+#pragma once
 #include "gmock/gmock.h"
 
 int main()
@@ -9,21 +9,14 @@ int main()
 }
 
 #else
-
+#pragma once
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
-
-#define CLEAR_SCREEN "\033[H\033[2J"
+#include "defineEnum.h"
+#include "DisplayMenu_Factory.h"
 
 int stack[10];
-
-void show_SelectMenu(const int& step);
-void printMenu_SelectCarType();
-void printMenu_SelectEngine();
-void printMenu_SelectBrake();
-void printMenu_SelectSteering();
-void printMenu_RunTest();
 
 void clear_UserInput(char  user_input[100]);
 bool isAnswer_Exit(const char* user_input);
@@ -31,71 +24,8 @@ bool check_AnswerIsNumber(char* user_input, int& answer);
 bool check_Invalid_inputNum(const int& step, const int& answer);
 
 void do_Answer(int& step, const int& answer);
-void selectCarType(const int& answer);
-void selectEngine(const int& answer);
-void selectbrakeSystem(const int& answer);
-void selectSteeringSystem(const int& answer);
 void runProducedCar();
 void testProducedCar();
-
-void delay(const int& ms);
-
-enum QuestionType
-{
-    CarType_Q,
-    Engine_Q,
-    brakeSystem_Q,
-    SteeringSystem_Q,
-    Run_Test_Q,
-};
-
-enum CarType
-{
-    SEDAN = 1,
-    SUV,
-    TRUCK
-};
-
-enum Engine
-{
-    GM = 1,
-    TOYOTA,
-    WIA
-};
-
-enum brakeSystem
-{
-    MANDO = 1,
-    CONTINENTAL,
-    BOSCH_B
-};
-
-enum SteeringSystem
-{
-    BOSCH_S = 1,
-    MOBIS
-};
-
-enum RunTest
-{
-    Run = 1,
-    Test
-};
-
-void delay(const int& ms)
-{
-    volatile int sum = 0;
-    for (int i = 0; i < 1000; i++)
-    {
-        for (int j = 0; j < 1000; j++)
-        {
-            for (int t = 0; t < ms; t++)
-            {
-                sum++;
-            }
-        }
-    }
-}
 
 int main()
 {
@@ -104,10 +34,9 @@ int main()
 
     while (1)
     {
-        show_SelectMenu(step);
-
-        printf("===============================\n");
-        printf("INPUT > ");
+        auto menu = DisplayMenu_Factory::createMenu(static_cast<QuestionType>(step));
+        menu->show_Menu();
+        menu->show_InputPrompt();
         fgets(user_input, sizeof(user_input), stdin);
 
         // 엔터 개행문자 제거
@@ -136,84 +65,9 @@ int main()
             continue;
         }
 
+        menu->show_YourSelect(answer);
         do_Answer(step, answer);
     }
-}
-
-
-void show_SelectMenu(const int& step)
-{
-    printf(CLEAR_SCREEN);
-    switch (step)
-    {
-    case CarType_Q:
-        printMenu_SelectCarType();
-        break;
-    case Engine_Q:
-        printMenu_SelectEngine();
-        break;
-    case brakeSystem_Q:
-        printMenu_SelectBrake();
-        break;
-    case SteeringSystem_Q:
-        printMenu_SelectSteering();
-        break;
-    case Run_Test_Q:
-        printMenu_RunTest();
-        break;
-    default:
-        break;
-    }
-}
-
-void printMenu_SelectCarType()
-{
-    printf("        ______________\n");
-    printf("       /|            | \n");
-    printf("  ____/_|_____________|____\n");
-    printf(" |                      O  |\n");
-    printf(" '-(@)----------------(@)--'\n");
-    printf("===============================\n");
-    printf("어떤 차량 타입을 선택할까요?\n");
-    printf("1. Sedan\n");
-    printf("2. SUV\n");
-    printf("3. Truck\n");
-}
-
-void printMenu_SelectEngine()
-{
-    printf("어떤 엔진을 탑재할까요?\n");
-    printf("0. 뒤로가기\n");
-    printf("1. GM\n");
-    printf("2. TOYOTA\n");
-    printf("3. WIA\n");
-    printf("4. 고장난 엔진\n");
-}
-
-void printMenu_SelectBrake()
-{
-    printf("어떤 제동장치를 선택할까요?\n");
-    printf("0. 뒤로가기\n");
-    printf("1. MANDO\n");
-    printf("2. CONTINENTAL\n");
-    printf("3. BOSCH\n");
-}
-
-void printMenu_SelectSteering()
-{
-    printf("어떤 조향장치를 선택할까요?\n");
-    printf("0. 뒤로가기\n");
-    printf("1. BOSCH\n");
-    printf("2. MOBIS\n");
-}
-
-void printMenu_RunTest()
-{
-    printf("멋진 차량이 완성되었습니다.\n");
-    printf("어떤 동작을 할까요?\n");
-    printf("0. 처음 화면으로 돌아가기\n");
-    printf("1. RUN\n");
-    printf("2. Test\n");
 }
 
 
@@ -296,80 +150,32 @@ void do_Answer(int& step, const int& answer)
     switch (step)
     {
     case CarType_Q:
-        selectCarType(answer);
+        stack[CarType_Q] = answer;
         step = Engine_Q;
         break;
     case Engine_Q:
-        selectEngine(answer);
+        stack[Engine_Q] = answer;
         step = brakeSystem_Q;
         break;
     case brakeSystem_Q:
-        selectbrakeSystem(answer);
+        stack[brakeSystem_Q] = answer;
         step = SteeringSystem_Q;
         break;
     case SteeringSystem_Q:
-        selectSteeringSystem(answer);
+        stack[SteeringSystem_Q] = answer;
         step = Run_Test_Q;
         break;
     case Run_Test_Q:
         if (answer == RunTest::Run)
             runProducedCar();
         else if (answer == RunTest::Test)
-        {
-            printf("Test...\n");
-            delay(1500);
             testProducedCar();
-        }
-        delayTime = 2000;
+        delay(2000);
         break;
     default:
         break;
     }
-    delay(delayTime);
 }
-
-void selectCarType(const int& answer)
-{
-    stack[CarType_Q] = answer;
-    if (answer == CarType::SEDAN)
-        printf("차량 타입으로 Sedan을 선택하셨습니다.\n");
-    if (answer == CarType::SUV)
-        printf("차량 타입으로 SUV을 선택하셨습니다.\n");
-    if (answer == CarType::TRUCK)
-        printf("차량 타입으로 Truck을 선택하셨습니다.\n");
-}
-
-void selectEngine(const int& answer)
-{
-    stack[Engine_Q] = answer;
-    if (answer == Engine::GM)
-        printf("GM 엔진을 선택하셨습니다.\n");
-    if (answer == Engine::TOYOTA)
-        printf("TOYOTA 엔진을 선택하셨습니다.\n");
-    if (answer == Engine::WIA)
-        printf("WIA 엔진을 선택하셨습니다.\n");
-}
-
-void selectbrakeSystem(const int& answer)
-{
-    stack[brakeSystem_Q] = answer;
-    if (answer == brakeSystem::MANDO)
-        printf("MANDO 제동장치를 선택하셨습니다.\n");
-    if (answer == brakeSystem::CONTINENTAL)
-        printf("CONTINENTAL 제동장치를 선택하셨습니다.\n");
-    if (answer == brakeSystem::BOSCH_B)
-        printf("BOSCH 제동장치를 선택하셨습니다.\n");
-}
-
-void selectSteeringSystem(const int& answer)
-{
-    stack[SteeringSystem_Q] = answer;
-    if (answer == SteeringSystem::BOSCH_S)
-        printf("BOSCH 조향장치를 선택하셨습니다.\n");
-    if (answer == SteeringSystem::MOBIS)
-        printf("MOBIS 조향장치를 선택하셨습니다.\n");
-}
-
 int isValidCheck()
 {
     if (stack[brakeSystem_Q] == BOSCH_B && stack[SteeringSystem_Q] != BOSCH_S)
