+++ mission2/mission2/defineEnum.h
@@ -1,0 +1,58 @@
+#pragma once
+
+enum QuestionType
+{
+    CarType_Q,
+    Engine_Q,
+    brakeSystem_Q,
+    SteeringSystem_Q,
+    Run_Test_Q,
+};
+
+enum CarType
+{
+    SEDAN = 1,
+    SUV,
+    TRUCK
+};
+
+enum Engine
+{
+    GM = 1,
+    TOYOTA,
+    WIA
+};
+
+enum brakeSystem
+{
+    MANDO = 1,
+    CONTINENTAL,
+    BOSCH_B
+};
+
+enum SteeringSystem
+{
+    BOSCH_S = 1,
+    MOBIS
+};
+
+enum RunTest
+{
+    Run = 1,
+    Test
+};
+
+static void delay(const int& ms)
+{
+    volatile int sum = 0;
+    for (int i = 0; i < 1000; i++)
+    {
+        for (int j = 0; j < 1000; j++)
+        {
+            for (int t = 0; t < ms; t++)
+            {
+                sum++;
+            }
+        }
+    }
+};
